AMP-9380
git-svn-id: file:///svn/branches/AMP_1_15_RELEASE/amp@13405 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/aim/action/ViewNewAdvancedReport.java
@@ -106,7 +106,7 @@ public class ViewNewAdvancedReport extends Action {
 				request.setAttribute("recordsPerPage", Integer.MAX_VALUE);
 			}
 		}else{
-			request.setAttribute("recordsPerPage", new Integer(25));
+			request.setAttribute("recordsPerPage", new Integer(100));
 		}
